Update __init__.py
importing the SpeechToTextPlugin from the speech_to_text_plugin file. The __all__ variable is a list that defines the public interface of the module. Here, it's saying that when we import the module using from module import, only the objects in __all__ will be imported, which is SpeechToTextPlugin in this case.